Encoder revamp (#202)
### Summary:
The existing encoder was **not** streaming. It would error if it didn't have enough space, but there was no way to know how much space it would need. We couldn't just try to re-encode the frame again when more space was available, because header-encoding mutates the HPACK state.

Now we have a "partially" streaming encoder. Most frames are pre-encoded, and we stream them into the available buffers. Headers and Data frames, which may need to split their payload across multiple frames, know how to do so.

### Details
- `aws_h2_frame` is a proper base class:
  - Previously, each specific frame type had its own init()/encode()/clean_up().  Now it's heap allocated and has vtable for encode()/destroy()
- Simple frame types (most) make use of the `aws_h2_frame_prebuilt` type.
  - The entire frame is pre-encoded.
  - We stream the pre-encoded frame into the available space of an aws_io_message.
- Remove "continuation" `aws_h2_frame` type.
  - Instead, the "headers" and "push promise" `aws_h2_frame` types will split their payload across CONTINUATION frames if necessary during encoding.
- Remove "data" `aws_h2_frame` type.
  - Instead, the encoder has a special function that takes a body-stream and writes a DATA frame immediately.
  - There are lots of differences in how the connection deals with DATA vs every other frame type, so it didn't make sense for it to be a proper `aws_h2_frame` type.
- HPACK encoding will grow the output buffer if necessary.
  - This was the simpler than building a fully "streaming" encoder.
  - HEADERS frame encoding is therefore 2 stage.
    - First, do HPACK encoding to separate buffer.
    - Then, split that payload across multiple HEADERS/CONTINUATION frames as necessary.
- Moved logic for encoding the header-block into hpack.c. Previously a bunch of that logic was in h2_frames.c. All this code was changing anyway.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.