Automatically reject PUSH_PROMISE streams (#228)
Until we have a need for it, PUSH_PROMISE is not a fully supported feature. Therefore, promised streams are automatically rejected in a manner compliant with RFC-7540

Why, you ask, don't we simply send SETTINGS_ENABLE_PUSH=0 in the initial SETTINGS frame and call it a day? Because it's theoretically possible for a server to start sending PUSH_PROMISE frames in the initial response, before sending the ACK to the initial SETTINGS.