move h1_connection struct to private .h file (#273)
rename `h1_connection` -> `aws_h1_connection`

Doing this so it's easier to share code between h1_connection.c/h1_stream.c. This is part of a larger change, but it's so disruptive I'm breaking it out into its own PR. This is similar to how we do it for h2, so that's nice too.