Revamp task scheduling logic in h1 connection and stream (#274)
* Revamp h1 connection/stream tasks to work like h2's

ISSUE
There were bugs in h1's task scheduling logic when chunks were involved. The logic was complicated to begin with. Adding support for chunks broke it, and it certainly wasn't going to accommodate the more complicated stuff we have coming down the pipe.

SOLUTION:
Make it work like h2. It's still complicated, but it can accommodate lots more functionality, and it's nice whenever h1 and h2 can be similar.