random access set (#361)
move the awslabs/aws-c-common#883 here as private structure since currently the use case is limited, not worth to make it public. Original review and comments can be fount from awslabs/aws-c-common#883. If anyone in the future needs the data structure outside http, he can continue on the PR in common.