Proxy h2 (#402)
- add support for using proxy to setup http/2 connection
- right now only supports talking h1 with proxy and create the http/2 connection from the tunneling proxy. 
- Turns out google doesn't like `:authority` with `host`, I am totally okay with what ever the real world want. :)

### TODO:
The `AWS_TEST_HTTPS_H2_PROXY_URL` now has a proxy host that can talk http/2 with client. Used apache server with mod_proxy_http2. You can simply use that host for the proxy option and update the tls of proxy options to have h2 in the alpn string to create the h2 connection with proxy. However, I don't know what's the expected behavior for client to talk to a h2 proxy host. 

**`curl` doesn't seem supporting it.**

So, put it as TODO, but still have the host here incase we want to implement it in the future.