Only "data" frames affect websocket's read window now (#407)
Fix it so that a websocket with manual_window_management turned off will disable backpressure on its whole channel (previously, we kept backpressure on and just automatically maintained a constant window size).

Change it so that the websocket's read window is only affected by "data" frames (TEXT, BINARY, CONTINUATION) and not "control" frames (PING, PONG, CLOSE). The assumption here being that only "data" frames are generating the kind of traffic that the user is forwarding along and needs tight control over. This is similar to how HTTP/2 only counts "data" frames as affecting the window.

Add copious documentation, because this stuff is confusing