Breaking change to websocket API: setup callback can report body of response (#409)
**Issue:**
If the server rejected a websocket connection, there was no way to see the body of the response. The response body often has useful information explaining why the connection was rejected.

**Underlying cause:**
The websocket code would close the connection immediately after receiving headers, because that was the simplest way to do things.

**Description of changes:**
Change the setup callback signature, so that the body can be reported.

The setup code keeps the connection open while receiving a rejection, making a copy of the body, and ultimately reporting it in the setup callback.