Validate websocket handshake response (#410)
**Issue:**
The websocket client did not fully validate the server's response, as specified in [RFC-6455 Section 4.1](https://www.rfc-editor.org/rfc/rfc6455#section-4.1)

The client was doing 1/6 required steps (checking the status-code, but not checking any headers)

**Description of changes:**
Validate the "Upgrade", "Connection", and "Sec-WebSocket-Accept" headers.

Now we're doing 4/6 required steps. Still need to check "Sec-WebSocket-Extensions" and "Sec-WebSocket-Protocol". That will be in a followup PR.